Converting the existing directive into typescript
Typescript all complete, haven't scrubbed the app's modules off this
version yet, will do tomorrow.  Pagination is working, but if you add
something like:

```HTML
<ul>
<li paginate="doll in matryoshka | itemsPerPage: 10">{{doll}}
<ul>
<li ng-repeat="doll in doll.others"></li>
</ul>
</li>
</ul>

```

The inner ng-repeat is currently getting commented out, but not put back
afterwards in the link function:

```HTML
<ul>
<li paginate="doll in matryoshka | itemsPerPage: 10">{{doll}}
<ul>
<!-- ng-repeat="doll in doll.others" -->
</ul>
</li>
</ul>

```

This seems to be an error in my typescript translation.  I modified the
directives to make them classes, so it's possible the link function
isn't getting called at the right time anymore.  Am continuing to search
for and fix the cause of this.  It functions as expected in the
[original
implementaion.](http://plnkr.co/edit/I0KWzX5lRR9CblwfSMTl?p=preview)